Revert "Eliminate make_map in favor of a simpler map property."
This reverts commit ca61e30d8f2b2a794374ab3f424df9d5337e778a.

As much as I like using the property instead of the function, it's way
more inefficient to generate the map every time it's accessed. Plus,
you need to seed the map with the controller list, which obviously
wouldn't stick.

I'd switch to using a memoization function of some sort or another,
but Python seems to have utterly failed to standardize on one.